Moved Path logics into separate files
